Remove CallbackRegistry
